+++ miles/utils/chat_template_utils/__init__.py
@@ -1,6 +1,5 @@
 """Chat template utilities for agentic-workflow token consistency."""
 
-from miles.utils.chat_template_utils.encoder import resolve_chat_template_encoder
 from miles.utils.chat_template_utils.template import (
     apply_chat_template,
     apply_chat_template_from_str,
@@ -32,7 +31,6 @@ __all__ = [
     "assert_messages_append_only_with_allowed_role",
     "message_matches",
     "extract_tool_dicts",
-    "resolve_chat_template_encoder",
     "Mismatch",
     "TokenSeqComparator",
     "MismatchType",
