+++ miles/utils/chat_template_utils/template.py
@@ -25,6 +25,8 @@ from pydantic import TypeAdapter
 from sglang.srt.entrypoints.openai.protocol import Tool
 from transformers.utils.chat_template_utils import render_jinja_template
 
+from miles.utils.chat_template_utils import deepseek_v32
+
 
 def load_hf_chat_template(model_id: str) -> str:
     """Load an original chat template from HuggingFace (cached locally).
@@ -221,6 +223,10 @@ def apply_chat_template(
     ensuring the result is ``str`` (tokenize=False) or ``list[int]``
     (tokenize=True), not a ``BatchEncoding`` or ``dict``.
     """
+    if deepseek_v32.is_deepseek_v32(tokenizer):
+        rendered = deepseek_v32.render_messages(messages, tools=tools, **kwargs)
+        return tokenizer.encode(rendered, add_special_tokens=False) if tokenize else rendered
+
     messages = _normalize_tool_arguments(messages)
     tool_defs = extract_tool_dicts(tools)
     render_kwargs = dict(add_generation_prompt=add_generation_prompt, **kwargs)
